refactor: extract insight merchant market map
