+++ frontend/src/insights/components/MerchantDistributionCard.tsx
@@ -1,34 +1,18 @@
-import {
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  type MouseEvent as ReactMouseEvent,
-  type TransitionEvent as ReactTransitionEvent,
-} from 'react'
+import { useMemo } from 'react'
 import { Store } from 'lucide-react'
 import type { FxStatus } from '@/api/shared/fx'
-import { ChartTooltipRow, ChartTooltipTitle } from '@/components/charts/ChartTooltipContent'
-import CursorTooltipPortal from '@/components/charts/CursorTooltipPortal'
 import IconTooltip from '@/components/IconTooltip'
-import type {
-  MerchantMarketMerchant,
-  MerchantMarketTile,
-} from '@/insights/types/merchantDistribution'
+import type { MerchantMarketMerchant } from '@/insights/types/merchantDistribution'
 import { getMerchantSpendingFxStatusMessage } from '@/insights/utils/fxTooltipMessages'
-import { formatCurrency } from '@/utils/formatCurrency'
-import { applyCursorTooltipPosition } from '@/utils/tooltipPosition'
 import {
   InsightLoadingContent,
   InsightLoadingOverlay,
 } from './InsightLoadingTransition'
 import { FxStatusBadge } from './FxStatusBadge'
+import { MerchantDistributionLegend } from './merchant-distribution/MerchantDistributionLegend'
+import { MerchantMarketMap } from './merchant-distribution/MerchantMarketMap'
 import { SectionHeader } from './SectionHeader'
 import { useInsightLoadingSnapshot } from './useInsightLoadingSnapshot'
-
-type MerchantMarketHover = {
-  merchant: MerchantMarketTile
-}
 
 type MerchantDistributionCardProps = {
   merchants: MerchantMarketMerchant[]
@@ -45,308 +29,9 @@ type MerchantDistributionSnapshot = {
   emptyLabel: string
 }
 
-const DEFAULT_MAP_WIDTH = 1000
-const DEFAULT_MAP_HEIGHT = 460
-const TOOLTIP_MARGIN = 8
-const TOOLTIP_PREFERRED_WIDTH = 300
-
-function formatSignedCurrency(amount: number, currency: string) {
-  if (amount === 0) return formatCurrency(amount, currency)
-  return `${amount > 0 ? '+' : '-'}${formatCurrency(Math.abs(amount), currency)}`
-}
-
-function splitTreemapItems(
-  items: MerchantMarketMerchant[],
-  x: number,
-  y: number,
-  width: number,
-  height: number,
-): MerchantMarketTile[] {
-  if (items.length <= 1) {
-    const item = items[0]
-    return item ? [{ ...item, x, y, width, height }] : []
-  }
-
-  const total = items.reduce((sum, item) => sum + item.totalAmount, 0)
-  let running = 0
-  let splitIndex = 1
-  for (let index = 0; index < items.length - 1; index += 1) {
-    running += items[index].totalAmount
-    if (running >= total / 2) {
-      splitIndex = index + 1
-      break
-    }
-  }
-
-  const firstGroup = items.slice(0, splitIndex)
-  const secondGroup = items.slice(splitIndex)
-  const firstTotal = firstGroup.reduce((sum, item) => sum + item.totalAmount, 0)
-  const firstShare = total > 0 ? firstTotal / total : 0.5
-
-  if (width >= height) {
-    const firstWidth = width * firstShare
-    return [
-      ...splitTreemapItems(firstGroup, x, y, firstWidth, height),
-      ...splitTreemapItems(secondGroup, x + firstWidth, y, width - firstWidth, height),
-    ]
-  }
-
-  const firstHeight = height * firstShare
-  return [
-    ...splitTreemapItems(firstGroup, x, y, width, firstHeight),
-    ...splitTreemapItems(secondGroup, x, y + firstHeight, width, height - firstHeight),
-  ]
-}
-
-function getMerchantMarketColor(changePct: number | null, changeAmount: number | null) {
-  if (changePct === null && changeAmount === null) {
-    return 'color-mix(in srgb, var(--app-text-muted) 24%, var(--app-input-bg))'
-  }
-  if (changePct === 0 || changeAmount === 0) {
-    return 'color-mix(in srgb, var(--app-accent) 14%, var(--app-input-bg))'
-  }
-  const direction = changePct ?? changeAmount ?? 0
-  const variable = direction < 0 ? 'var(--app-chart-positive)' : 'var(--app-chart-negative)'
-  const mix = changePct === null ? 34 : Math.min(72, 24 + Math.abs(changePct) * 2.2)
-  return `color-mix(in srgb, ${variable} ${mix}%, var(--app-input-bg))`
-}
-
-function getMerchantTileColor(merchant: MerchantMarketTile) {
-  if (merchant.id === 'other-merchants') {
-    return 'color-mix(in srgb, var(--app-text-muted) 24%, var(--app-input-bg))'
-  }
-  return getMerchantMarketColor(merchant.changePct, merchant.changeAmount)
-}
-
-function MerchantDistributionLegend({ className = '' }: { className?: string }) {
-  return (
-    <div className={`flex flex-wrap items-center gap-3 ${className}`} style={{ color: 'var(--app-text-muted)' }}>
-      <span className="inline-flex items-center gap-1.5">
-        <span className="h-2.5 w-2.5 rounded-sm" style={{ background: 'var(--app-chart-positive)' }} />
-        Spend down
-      </span>
-      <span className="inline-flex items-center gap-1.5">
-        <span className="h-2.5 w-2.5 rounded-sm" style={{ background: 'var(--app-chart-negative)' }} />
-        Spend up
-      </span>
-      <span className="inline-flex items-center gap-1.5">
-        <span className="h-2.5 w-2.5 rounded-sm" style={{ background: 'var(--app-accent)' }} />
-        Flat
-      </span>
-    </div>
-  )
-}
-
-function MerchantMarketMap({
-  merchants,
-  currency,
-}: {
-  merchants: MerchantMarketMerchant[]
-  currency: string
-}) {
-  const [hoveredTile, setHoveredTile] = useState<MerchantMarketHover | null>(null)
-  const mapRef = useRef<HTMLDivElement | null>(null)
-  const tooltipRef = useRef<HTMLDivElement | null>(null)
-  const [tooltipVisible, setTooltipVisible] = useState(false)
-  const [mapSize, setMapSize] = useState({ width: DEFAULT_MAP_WIDTH, height: DEFAULT_MAP_HEIGHT })
-  const layoutWidth = Math.max(mapSize.width, 1)
-  const layoutHeight = Math.max(mapSize.height, 1)
-  const tooltipMaxWidth = Math.max(layoutWidth - TOOLTIP_MARGIN * 2, 1)
-  const tooltipWidth = Math.min(TOOLTIP_PREFERRED_WIDTH, tooltipMaxWidth)
-  const laidOutMerchants = useMemo(
-    () => splitTreemapItems(merchants, 0, 0, layoutWidth, layoutHeight),
-    [layoutHeight, layoutWidth, merchants],
-  )
-  const updateMerchantTooltipPosition = (clientX: number, clientY: number) => {
-    const map = mapRef.current
-    const tooltip = tooltipRef.current
-    if (!map || !tooltip) return
-
-    applyCursorTooltipPosition({
-      origin: map,
-      tooltip,
-      clientX,
-      clientY,
-      xProperty: '--merchant-tooltip-x',
-      yProperty: '--merchant-tooltip-y',
-    })
-  }
-  const showMerchantTooltip = (
-    merchant: MerchantMarketTile,
-    event: ReactMouseEvent<Element>,
-  ) => {
-    updateMerchantTooltipPosition(event.clientX, event.clientY)
-    setHoveredTile((current) => (current?.merchant.id === merchant.id ? current : { merchant }))
-    setTooltipVisible(true)
-    requestAnimationFrame(() => updateMerchantTooltipPosition(event.clientX, event.clientY))
-  }
-  const hideMerchantTooltip = () => {
-    setTooltipVisible(false)
-  }
-  const handleMerchantTooltipTransitionEnd = (event: ReactTransitionEvent<HTMLDivElement>) => {
-    if (event.target !== event.currentTarget || event.propertyName !== 'opacity' || tooltipVisible) return
-    setHoveredTile(null)
-  }
-
-  useEffect(() => {
-    const element = mapRef.current
-    if (!element || typeof ResizeObserver === 'undefined') return undefined
-
-    const updateSize = (width: number, height: number) => {
-      const nextWidth = Math.max(Math.round(width), 1)
-      const nextHeight = Math.max(Math.round(height), 1)
-      setMapSize((current) => {
-        if (current.width === nextWidth && current.height === nextHeight) return current
-        return { width: nextWidth, height: nextHeight }
-      })
-    }
-
-    const rect = element.getBoundingClientRect()
-    updateSize(rect.width, rect.height)
-
-    const observer = new ResizeObserver(([entry]) => {
-      updateSize(entry.contentRect.width, entry.contentRect.height)
-    })
-    observer.observe(element)
-    return () => observer.disconnect()
-  }, [])
-
-  return (
-    <div
-      ref={mapRef}
-      className="relative min-h-0 flex-1"
-      onMouseLeave={hideMerchantTooltip}
-    >
-      <div className="absolute inset-0 overflow-hidden rounded-lg border border-[var(--app-border)]">
-        <svg
-          viewBox={`0 0 ${layoutWidth} ${layoutHeight}`}
-          preserveAspectRatio="none"
-          role="img"
-          aria-label="Merchant market map"
-          className="pointer-events-none h-full w-full"
-        >
-          {laidOutMerchants.map((merchant) => (
-            <rect
-              key={merchant.id}
-              x={merchant.x + 2}
-              y={merchant.y + 2}
-              width={Math.max(merchant.width - 4, 0)}
-              height={Math.max(merchant.height - 4, 0)}
-              rx={6}
-              fill={getMerchantTileColor(merchant)}
-              stroke="var(--app-surface-soft)"
-              strokeWidth={4}
-            />
-          ))}
-        </svg>
-        <div className="absolute inset-0">
-          {laidOutMerchants.map((merchant) => {
-            const area = merchant.width * merchant.height
-            const labelSize = area > 110000 ? 20 : area > 42000 ? 17 : 15
-            const amountSize = Math.max(labelSize - 3, 12)
-            const amountText = formatCurrency(merchant.totalAmount, currency)
-            const contentWidth = Math.max(merchant.width - 20, 0)
-            const contentHeight = Math.max(merchant.height - 20, 0)
-            const showName = contentWidth >= 80 && contentHeight >= 42
-            const showAmount = contentWidth >= 70 && contentHeight >= (showName ? 62 : 30)
-            const compactName = contentWidth < 128 || contentHeight < 74
-            const showDetailsIndicator = !showName && !showAmount && contentWidth >= 18 && contentHeight >= 18
-            const detailsIndicatorSize = Math.min(20, Math.max(15, Math.min(contentWidth, contentHeight) * 0.58))
-            return (
-              <div
-                key={merchant.id}
-                className="absolute flex min-w-0 cursor-default flex-col items-center justify-center overflow-hidden p-2 text-center"
-                style={{
-                  color: 'var(--app-text)',
-                  left: `${(merchant.x / layoutWidth) * 100}%`,
-                  top: `${(merchant.y / layoutHeight) * 100}%`,
-                  width: `${(merchant.width / layoutWidth) * 100}%`,
-                  height: `${(merchant.height / layoutHeight) * 100}%`,
-                }}
-                onMouseEnter={(event) => showMerchantTooltip(merchant, event)}
-                onMouseMove={(event) => showMerchantTooltip(merchant, event)}
-              >
-                {showName && (
-                  <p
-                    className="max-w-full overflow-hidden font-bold leading-tight"
-                    style={{
-                      fontSize: labelSize,
-                      overflowWrap: 'normal',
-                      textOverflow: 'ellipsis',
-                      whiteSpace: compactName ? 'nowrap' : undefined,
-                      wordBreak: 'normal',
-                      ...(!compactName
-                        ? {
-                            display: '-webkit-box',
-                            WebkitBoxOrient: 'vertical',
-                            WebkitLineClamp: 2,
-                          }
-                        : {}),
-                    }}
-                  >
-                    {merchant.name}
-                  </p>
-                )}
-                {showAmount && (
-                  <p
-                    className={`${showName ? 'mt-1' : ''} max-w-full truncate font-financial leading-tight`}
-                    style={{ color: 'var(--app-text-muted)', fontSize: amountSize }}
-                  >
-                    {amountText}
-                  </p>
-                )}
-                {showDetailsIndicator && (
-                  <span
-                    aria-hidden
-                    className="font-bold leading-none"
-                    style={{ color: 'var(--app-text-muted)', fontSize: detailsIndicatorSize }}
-                  >
-                    ...
-                  </span>
-                )}
-              </div>
-            )
-          })}
-        </div>
-      </div>
-      <CursorTooltipPortal
-        ref={tooltipRef}
-        onTransitionEnd={handleMerchantTooltipTransitionEnd}
-        style={{
-          opacity: tooltipVisible ? 1 : 0,
-          transform: 'translate3d(var(--merchant-tooltip-x, 0px), var(--merchant-tooltip-y, 0px), 0)',
-          width: tooltipWidth,
-        }}
-      >
-        {hoveredTile && (
-          <>
-            <ChartTooltipTitle>{hoveredTile.merchant.name}</ChartTooltipTitle>
-            <ChartTooltipRow
-              label="Total Spend"
-              value={formatCurrency(hoveredTile.merchant.totalAmount, currency)}
-              financialValue
-            />
-            {hoveredTile.merchant.changeAmount !== null && (
-              <ChartTooltipRow
-                label="Change"
-                value={(
-                  <>
-                    {formatSignedCurrency(hoveredTile.merchant.changeAmount, currency)}
-                    {hoveredTile.merchant.changePct === null
-                      ? ''
-                      : ` (${hoveredTile.merchant.changePct > 0 ? '+' : ''}${hoveredTile.merchant.changePct}%)`}
-                  </>
-                )}
-                financialValue
-              />
-            )}
-          </>
-        )}
-      </CursorTooltipPortal>
-    </div>
-  )
-}
-
+/**
+ * Renders merchant distribution totals and the merchant market map
+ */
 export function MerchantDistributionCard({
   merchants,
   fxStatus,
